move replicate_all to rereplicate
